# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quant Lab Analysis Service API
  version: 1.0.0
  description: |
    A-share quantitative trading platform REST API.

    The Analysis Service acts as the API gateway, coordinating backtest,
    strategy, data, risk, execution, compliance, and AI research endpoints.

    ## Authentication
    Most endpoints are open in dev mode. When `auth.jwt_secret` is configured,
    protected endpoints require a `Bearer` JWT token obtained via `POST /api/auth/login`.

    ## Conventions
    - All dates use `YYYY-MM-DD` unless noted otherwise.
    - Monetary values are in CNY (Chinese Yuan).
    - Symbols follow the `XXXXXX.SH` / `XXXXXX.SZ` convention.
  contact:
    name: Quant Lab
    url: https://github.com/ruoxizhnya/quant-trading
  license:
    name: MIT
servers:
  - url: http://localhost:8085
    description: Local development server
  - url: http://analysis-service:8085
    description: Docker compose internal URL
...
paths:
...
  /api/execution/orders:
    post:
      tags: [Execution]
      summary: Submit an order
      description: |
        Submits a paper-trading order to the in-process MockTrader.

        AUD-02 (ODR-065 H5): requires role `trader` or `admin`. A `viewer`
        token gets 403. The legacy root path `POST /orders` carries the
        same requirement — the two are the same handler on two paths.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid access token
        '403':
          description: Authenticated but role is not trader/admin
    get:
      tags: [Execution]
      summary: List orders
      operationId: listOrders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResult'
                  count:
                    type: integer
  /api/execution/orders/{id}:
    get:
      tags: [Execution]
      summary: Get an order by ID
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/execution/orders/{id}/cancel:
    post:
      tags: [Execution]
      summary: Cancel an order
      description: |
        AUD-02 (ODR-065 H5): requires role `trader` or `admin`. The legacy
        root path `POST /orders/{id}/cancel` carries the same requirement.
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: string
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid access token
        '403':
          description: Authenticated but role is not trader/admin
...
  /api/execution/emergency-flatten:
    post:
      tags: [Execution]
      summary: Emergency flatten (kill-switch)
      description: |
        Liquidates all positions immediately. Requires:
        1. Server-side `trading.emergency_token` configured.
        2. Bearer token in Authorization header matching the server token.
        3. JSON body with `reason` and `confirmation_token`.
      operationId: emergencyFlatten
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, confirmation_token]
              properties:
                reason:
                  type: string
                confirmation_token:
                  type: string
      responses:
        '200':
          description: Flatten result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyFlattenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Invalid token or confirmation mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Kill-switch disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================
  # Compliance
  # ============================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained via POST /api/auth/login.

  # ============================================================
  # Parameters
  # ============================================================
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        status:
          type: string
...
    CreateOrderRequest:
      type: object
      required: [symbol, side, quantity]
      properties:
        symbol:
          type: string
          example: "000001.SZ"
        side:
          type: string
          enum: [long, short, close]
        type:
          type: string
          enum: [market, limit, stop, trailing]
          default: market
        quantity:
          type: number
          minimum: 0
        price:
          type: number
    OrderResult:
      type: object
      properties:
        order_id:
          type: string
        symbol:
          type: string
        status:
          type: string
        filled_qty:
          type: number
        avg_price:
          type: number
        commission:
          type: number
        timestamp:
          type: string
          format: date-time
    EmergencyFlattenResponse:
      type: object
      properties:
        sold:
          type: array
          items:
            type: object
        skipped:
          type: array
          items:
            type: object
        sold_total:
          type: number
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        reason:
          type: string
        latency_ms:
          type: integer
...